Chore, Refactor: RCPT 2.6.0-alpha, JAP error-type reporting.
Bump room-compiler-processing-testing to 2.6.0-alpha
 which includes `correctErrorTypes` fix for testing.
 See #24 for more info.

Remove google-auto `BasicAnnotationProcessor` harness, as
 all it does is reports unresolved types, which yatagan reports
 by its own means.

Theoretically, this completely removes support for multi-round
 processing. In practice the support was broken from the very beginning
 and yatagan doesn't intend to support it
 as it slows the build and, obviously, doesn't make sense to use
 with reflection backend.

Fix the `invalid-component-creator` test as it contained
 non-deterministic behavior (the errors reported depended on
 the inner class declaration order, which currently is not forced
 to be the same across backends).
