+++ processor/jap/build.gradle.kts
@@ -3,8 +3,6 @@ plugins {
 }
 
 dependencies {
-    api(libs.autoCommon)
-
     implementation(project(":api:public"))
     implementation(project(":processor:common"))
     implementation(project(":lang:jap"))
