+++ processor/jap/src/main/kotlin/JapLogger.kt
@@ -21,7 +21,7 @@ import javax.annotation.processing.Messager
 import javax.tools.Diagnostic.Kind.ERROR
 import javax.tools.Diagnostic.Kind.MANDATORY_WARNING
 
-class JapLogger(private val messager: Messager) : Logger {
+internal class JapLogger(private val messager: Messager) : Logger {
     override fun error(message: String) {
         messager.printMessage(ERROR, message)
     }
